Add tests for incorrect treatment of RawString and introduce unwrap_raw
To allow for people to use RawStrings as normal Strings when they really really
want to (for example in StringConcatenation, or with an Extension that doesn't
support RawString), this commit introduces the unwrap_raw filter, which has the
inverse effect of raw.

Additionally, test cases have been introduced that check the behaviour of this.